pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "fairway"
version = "0.2.0"
authors = [
  { name="Antigravity", email="antigravity@google.com" },
]
description = "A portable data ingestion framework for centralized research data."
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "click",
    "pyyaml",
    "pandas",
    "tabulate",
    "platformdirs>=4.0",
    "python-ulid>=2.0",
]

[project.optional-dependencies]
spark = [
    "pyspark==4.1.1",
    "delta-spark>=3.0.0", # Added for Option B support
    "pandas",
    "numpy",
    "pyarrow",
]
duckdb = [
    "duckdb",
    "numpy",
    "pyarrow",
]
test-data-gen = [
    "numpy",
]
all = [
    "pyspark>=4.1.1",
    "delta-spark>=3.0.0",
    "duckdb",
    "numpy",
    "pyarrow",
]
dev = [
    "pytest",
    "pytest-cov",
]
docs = [
    "mkdocs-material",
    "mkdocs-coverage",
]

[project.scripts]
fairway = "fairway.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools]
include-package-data = true

[tool.pytest.ini_options]
markers = [
    "hpc: requires SLURM cluster — skipped in local and container runs",
    "slow: takes >10 seconds",
    "local: runs only in local (non-cluster) environments",
    "spark: requires PySpark — skipped when Spark is unavailable",
    "no_fairway_home: opt out of the autouse FAIRWAY_HOME/FAIRWAY_SCRATCH fixture (for resolver-behavior tests)",
]
testpaths = ["tests"]
addopts = "--cov=src/fairway --cov-report=term-missing --cov-report=xml:build/coverage/coverage.xml --cov-report=html:build/coverage/html --cov-fail-under=67 --basetemp=build/test-tmp"

[tool.coverage.run]
data_file = "build/coverage/.coverage"
relative_files = true

[tool.setuptools.package-data]
"fairway.data" = ["*.yaml", "*.md", "*.txt", "*.py", "*.def", "Dockerfile", "Makefile", ".dockerignore"]
"fairway.data.scripts" = ["*.sh", "*.py"]
"fairway.slurm_templates" = ["*.sh"]
